% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-aci.R
\name{calc_aci}
\alias{calc_aci}
\title{Calculate ACI measure}
\usage{
calc_aci(pop, popsh, est, se, SEuseful, subgroup_order, ...)
}
\arguments{
\item{pop}{Pop.}

\item{popsh}{Popsh.}

\item{est}{Estimate.}

\item{se}{Se.}

\item{SEuseful}{A boolean?}

\item{...}{Other arguments}
}
\value{
A tibble.
}
\description{
Calculate ACI measure
}
