% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-bgsd.R
\name{calc_bgsd}
\alias{calc_bgsd}
\title{Caculate BGSD}
\usage{
calc_bgsd(popsh, est, est_natl, se, SEuseful, ...)
}
\arguments{
\item{popsh}{Popsh}

\item{est}{Est}

\item{est_natl}{Est national}

\item{se}{Standard}

\item{SEuseful}{A boolean}

\item{...}{Other arguments.}
}
\value{
A tibble.
}
\description{
Caculate BGSD
}
