% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-bgv.R
\name{calc_bgv}
\alias{calc_bgv}
\title{Calculate between-groups variance}
\usage{
calc_bgv(popsh, est, est_natl, se, SEuseful, ...)
}
\arguments{
\item{popsh}{Population share.}

\item{est}{Estimate.}

\item{est_natl}{National estimate.}

\item{se}{Standard}

\item{SEuseful}{}

\item{...}{Other arguments.}
}
\value{
A tibble.
}
\description{
Calculate between-groups variance
}
