% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-cov.R
\name{calc_cov}
\alias{calc_cov}
\title{Title}
\usage{
calc_cov(popsh, est, est_natl, se, SEuseful, ...)
}
\arguments{
\item{popsh}{}

\item{est}{}

\item{est_natl}{}

\item{se}{}

\item{SEuseful}{}

\item{...}{}
}
\value{

}
\description{
Title
}
