% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-d.R
\name{calc_d}
\alias{calc_d}
\title{Caculate D}
\usage{
calc_d(max_est, min_est, max_se, min_se, ...)
}
\arguments{
\item{max_est}{Max.}

\item{min_est}{Min.}

\item{max_se}{Max.}

\item{min_se}{Min.}

\item{...}{}
}
\value{
A tibble.
}
\description{
Caculate D
}
