% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-idisw.R
\name{calc_idisw}
\alias{calc_idisw}
\title{Title}
\usage{
calc_idisw(popsh, est, est_natl, se, SEuseful, scaleval, simulations, ...)
}
\arguments{
\item{popsh}{}

\item{est}{}

\item{est_natl}{}

\item{se}{}

\item{SEuseful}{}

\item{scaleval}{}

\item{simulations}{}

\item{...}{}
}
\value{

}
\description{
Title
}
