% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-mdbw.R
\name{calc_mdbw}
\alias{calc_mdbw}
\title{Title}
\usage{
calc_mdbw(
  popsh,
  est,
  se,
  SEuseful,
  favourable,
  hasrefgroup,
  scaleval,
  simulations,
  ...
)
}
\arguments{
\item{popsh}{}

\item{est}{}

\item{se}{}

\item{SEuseful}{}

\item{favourable}{}

\item{hasrefgroup}{}

\item{scaleval}{}

\item{simulations}{}

\item{...}{}
}
\value{

}
\description{
Title
}
