% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-mdmu.R
\name{calc_mdmu}
\alias{calc_mdmu}
\title{Title}
\usage{
calc_mdmu(pop, est, se, SEuseful, scaleval, simulations, ...)
}
\arguments{
\item{pop}{}

\item{est}{}

\item{se}{}

\item{SEuseful}{}

\item{scaleval}{}

\item{simulations}{}

\item{...}{}
}
\value{

}
\description{
Title
}
