% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-mdmw.R
\name{calc_mdmw}
\alias{calc_mdmw}
\title{Title}
\usage{
calc_mdmw(popsh, est, se, SEuseful, scaleval, simulations, ...)
}
\arguments{
\item{popsh}{}

\item{est}{}

\item{se}{}

\item{SEuseful}{}

\item{scaleval}{}

\item{simulations}{}

\item{...}{}
}
\value{

}
\description{
Title
}
