% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-mdru.R
\name{calc_mdru}
\alias{calc_mdru}
\title{Title}
\usage{
calc_mdru(
  est,
  se,
  SEuseful,
  favourable,
  scaleval,
  simulations,
  reference_subgroup,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
