% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-mdrw.R
\name{calc_mdrw}
\alias{calc_mdrw}
\title{Title}
\usage{
calc_mdrw(
  est,
  se,
  popsh,
  SEuseful,
  favourable,
  scaleval,
  simulations,
  reference_subgroup,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
