% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-par.R
\name{calc_par}
\alias{calc_par}
\title{Title}
\usage{
calc_par(
  pop,
  est,
  est_natl,
  se,
  SEuseful,
  ordered,
  favourable,
  hasrefgroup,
  reference_subgroup,
  max_est,
  min_est,
  subgroup_order,
  scaleval,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
