% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-r.R
\name{calc_r}
\alias{calc_r}
\title{Title}
\usage{
calc_r(ordered, favourable, max_est, min_est, max_se, min_se, ...)
}
\arguments{
\item{ordered}{}

\item{favourable}{}

\item{max_est}{}

\item{min_est}{}

\item{max_se}{}

\item{min_se}{}

\item{...}{}
}
\description{
Title
}
