% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-rci.R
\name{calc_rci}
\alias{calc_rci}
\title{Title}
\usage{
calc_rci(pop, popsh, est, est_natl, se, SEuseful, subgroup_order, ...)
}
\arguments{
\item{pop}{}

\item{popsh}{}

\item{est}{}

\item{est_natl}{}

\item{se}{}

\item{SEuseful}{}

\item{...}{}
}
\description{
Title
}
