% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-rii.R
\name{calc_rii}
\alias{calc_rii}
\title{Title}
\usage{
calc_rii(pop, est, se, SEuseful, favourable, scaleval, subgroup_order, ...)
}
\arguments{
\item{pop}{}

\item{est}{}

\item{se}{}

\item{SEuseful}{}

\item{favourable}{}

\item{scaleval}{}

\item{subgroup_order}{}

\item{...}{}
}
\value{

}
\description{
Title
}
