% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-sii.R
\name{calc_sii}
\alias{calc_sii}
\title{Slope Index of Inequality (SII)
relative, complex, ordered, weighted, greater than two (wealth, educ)}
\usage{
calc_sii(pop, est, se, SEuseful, favourable, scaleval, subgroup_order, ...)
}
\arguments{
\item{pop}{}

\item{est}{}

\item{se}{}

\item{SEuseful}{}

\item{favourable}{}

\item{scaleval}{}

\item{subgroup_order}{}

\item{...}{}
}
\value{

}
\description{
Slope Index of Inequality (SII)
relative, complex, ordered, weighted, greater than two (wealth, educ)
}
