% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_weighted_mean}
\alias{get_weighted_mean}
\title{See github 903}
\usage{
get_weighted_mean(est, popsh, setting_average)
}
\arguments{
\item{est}{}

\item{popsh}{}

\item{setting_average}{}
}
\description{
See github 903
}
