% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_inequality_measures.R
\name{HEAT_choose_measures_dataset}
\alias{HEAT_choose_measures_dataset}
\title{A function to identify which inequality measures to use for a full dataset based on dimensions}
\usage{
HEAT_choose_measures_dataset(.data)
}
\arguments{
\item{.data}{a full dataset, potentially with many dimensions}
}
\description{
A function to identify which inequality measures to use for a full dataset based on dimensions
}
