% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_inequality_measures.R
\name{HEAT_choose_measures_strata}
\alias{HEAT_choose_measures_strata}
\title{A function to identify which inequality measures to use for a strata}
\usage{
HEAT_choose_measures_strata(strata)
}
\arguments{
\item{strata}{one strata worth of data as a data.frame or tibble}
}
\value{

}
\description{
A function to identify which inequality measures to use for a strata
}
\examples{
filter(sample_strata, strata_id == sample_strata$strata_id[1]) \%>\% 
which_inequality_measures()
}
