% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{HEAT_create_input_list}
\alias{HEAT_create_input_list}
\title{Create list of inputs}
\usage{
HEAT_create_input_list(
  .data,
  validation_results,
  simulation_n = 100,
  force_simulations = FALSE
)
}
\arguments{
\item{.data}{A data frame.}

\item{force_simulations}{for testing if you want to get simulations even if
they wouldn't be calculated otherwise}
}
\value{
A list.
}
\description{
Create list of inputs
}
