% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{HEAT_measures_full_process}
\alias{HEAT_measures_full_process}
\title{Apply validation tests and compute inequality measures in one step}
\usage{
HEAT_measures_full_process(.data)
}
\arguments{
\item{.data}{main data as a data.frame or tibble}
}
\value{

}
\description{
Apply validation tests and compute inequality measures in one step
}
