% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{HEAT_measures_map}
\alias{HEAT_measures_map}
\title{Title}
\usage{
HEAT_measures_map(.data_split, validation_results)
}
\arguments{
\item{.data_split}{}

\item{validation_results}{}
}
\value{

}
\description{
Title
}
