% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{HEAT_measures_outlier2NA}
\alias{HEAT_measures_outlier2NA}
\title{Set outlying inequality measures to NA}
\usage{
HEAT_measures_outlier2NA(.data)
}
\arguments{
\item{.data}{}
}
\value{

}
\description{
Outliers are defined in the summary measures Excel file in data-raw. They are defined
solely on the value of inequal -- there is an inequal_min and inequal_max field. This function
identifies if inequal is in range between those and if not it sets inequal, se, se.lowerci
and se.upperci to NA.
}
