% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{HEAT_measures_validation_map}
\alias{HEAT_measures_validation_map}
\title{Applies validation tests to a set of strata}
\usage{
HEAT_measures_validation_map(.data_split, measures)
}
\arguments{
\item{.data_split}{a list created by breaking up a dataset by strata}

\item{measures}{the set of measures}
}
\value{
A list.
}
\description{
Applies validation tests to a set of strata
}
