% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{HEAT_measures_validation_tests}
\alias{HEAT_measures_validation_tests}
\title{Map the validation tests for inequality measures to a strata}
\usage{
HEAT_measures_validation_tests(strata, measures, only_return_passing = FALSE)
}
\arguments{
\item{strata}{a dataset representing data for a single strata}

\item{measures}{a vector of inequality measure abbreviations}

\item{only_return_passing}{if TRUE it only returns a vector of summary
measures that pass all tests otherwise returns a list (see below Value)}
}
\value{
If only_return_passing = TRUE it only returns a vector of summary measures
that pass all tests otherwise returns a list with three items. One is a
vector of those that pass \verb{Measures that pass}, those that fail \verb{Measures that fail} and a table of test results describing which tests fail.
}
\description{
Makes use of a built in table called summary_measure_types to determine the
appropriate test.
}
\details{
\preformatted{Observations: 31
Variables: 5
$ measure         <chr> "aci", "bgsd", "bgv", "cov", "d", "d", "d", "d", "idisu", "idisw", "mdbu", ... \\cr
$ just2_subgroups <lgl> FALSE, FALSE, FALSE, FALSE, TRUE, FALSE, FALSE, TRUE, FALSE, FALSE, FALSE, ... \\cr
$ ordered         <lgl> TRUE, FALSE, FALSE, FALSE, FALSE, FALSE, TRUE, TRUE, FALSE, FALSE, FALSE, F... \\cr
$ simulation_se   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, ... \\cr
$ test_name       <list> [<"test_estimates_all_zero", "test_two_subgroups", "test_less_than2_subgro... \\cr
}
}
