% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmeasures.R
\docType{package}
\name{heatmeasures-package}
\alias{heatmeasures}
\alias{heatmeasures-package}
\title{HEAT Measures}
\description{
HEAT Measures application
}
\author{
\strong{Maintainer}:  ZevRoss Spatial Anaylsis \email{info@zevross.com}

Other contributors:
\itemize{
  \item Zev Ross [contributor]
}

}
