% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{midPointProp}
\alias{midPointProp}
\title{Returns the cumulative mid point proportion of each group}
\usage{
midPointProp(w)
}
\arguments{
\item{w}{a vector of numbers of the population in each group}
}
\value{
A vector representing the cumulative mid-point proportion
}
\description{
Returns the cumulative mid point proportion of each group
}
