% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-paf.R
\name{calc_paf}
\alias{calc_paf}
\title{Title}
\usage{
calc_paf(
  pop,
  est,
  est_natl,
  se,
  SEuseful,
  ordered,
  favourable,
  hasrefgroup,
  reference_subgroup,
  max_est,
  min_est,
  subgroup_order,
  scaleval,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
