% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-rii.R
\name{calc_rii}
\alias{calc_rii}
\title{Slope Index of Inequality (SII)
relative, complex, ordered, weighted, greater than two (wealth, educ)}
\usage{
calc_rii(pop, est, se, SEuseful, favourable, scaleval, subgroup_order, ...)
}
\arguments{
\item{...}{}
}
\description{
Slope Index of Inequality (SII)
relative, complex, ordered, weighted, greater than two (wealth, educ)
}
